packages/py-common/pyproject.toml
[project]
name = "py-common"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "structlog>=24.1.0",
  "nats-py>=2.7.0",
  "psycopg[binary]>=3.1.18",
  "psycopg-pool>=3.2.0",
  "boto3>=1.34.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
  "jsonschema>=4.21.0"
]
[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.3.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["py_common*"]
